Reorganise TaskHelpField
